+++ ssr/src/page/mod.rs
@@ -10,6 +10,8 @@ pub mod logout;
 pub mod menu;
 pub mod notifs;
 pub mod post_view;
+#[cfg(any(feature = "oauth-ssr", feature = "oauth-hydrate"))]
+pub mod preview_google_redirect;
 pub mod privacy;
 pub mod profile;
 pub mod refer_earn;
